+++ src/NuGet.Core/NuGet.Build.Tasks.Pack/PackTask.cs
@@ -9,7 +9,6 @@ using NuGet.Common;
 using NuGet.Packaging;
 using ILogger = NuGet.Common.ILogger;
 using PackageSpecificWarningProperties = NuGet.Commands.PackCommand.PackageSpecificWarningProperties;
-using WarningPropertiesCollection = NuGet.Commands.PackCommand.WarningPropertiesCollection;
 
 namespace NuGet.Build.Tasks.Pack
 {
@@ -117,7 +116,7 @@ namespace NuGet.Build.Tasks.Pack
                 var request = GetRequest();
                 var logic = PackTaskLogic;
                 PackageBuilder packageBuilder = null;
-                var packArgs = logic.GetPackArgs(request: request, warningPropertiesCollection: null);
+                var packArgs = logic.GetPackArgs(request: request, packageSpecificWarningProperties: null);
 
                 // If packing using a Nuspec file, we don't need to build a PackageBuilder here
                 // as the package builder is built by reading the manifest file later in the code path.
@@ -128,10 +127,9 @@ namespace NuGet.Build.Tasks.Pack
 
                     if (packageBuilder?.PackageSpecificNoWarnProperties.Keys.Count > 0)
                     {
-                        var warningPropertiesCollection = new WarningPropertiesCollection(PackageSpecificWarningProperties
-                            .CreatePackageSpecificWarningProperties(packageBuilder.PackageSpecificNoWarnProperties));
+                        PackageSpecificWarningProperties packageSpecificWarningProperties = PackageSpecificWarningProperties.CreatePackageSpecificWarningProperties(packageBuilder.PackageSpecificNoWarnProperties);
                         // Override logger with warning properties
-                        packArgs = logic.GetPackArgs(request, warningPropertiesCollection);
+                        packArgs = logic.GetPackArgs(request, packageSpecificWarningProperties);
                     }
                 }
