+++ test/NuGet.Core.FuncTests/Dotnet.Integration.Test/PackCommandTests.cs
@@ -5958,17 +5958,7 @@ namespace ClassLibrary
                 settings.AddNetStandardFeeds();
 
                 string testDirectory = pathContext.WorkingDirectory;
-                var customNuGetConfigContent = @"<configuration>
-  <packageSources>
-    <clear />
-    <add key='nuget' value ='https://api.nuget.org/v3/index.json' />
-    <add key ='local' value ='../pkgs' />
-  </packageSources>
-</configuration>";
-
-                File.WriteAllText(Path.Combine(testDirectory, "NuGet.Config"), customNuGetConfigContent);
-                var mainDirectory = Directory.GetParent(testDirectory).FullName;
-                string packageSource = Path.Combine(mainDirectory, "pkgs");
+                string packageSource = pathContext.PackageSource;
 
                 var prereleaseDependencyAName = "PreReleasePackageA";
                 var prereleaseDependencyAVersion = "4.8.0-beta00011";
@@ -6028,8 +6018,7 @@ namespace ClassLibrary
                 settings.AddNetStandardFeeds();
 
                 string testDirectory = pathContext.WorkingDirectory;
-                var mainDirectory = Directory.GetParent(testDirectory).FullName;
-                string packageSource = Path.Combine(mainDirectory, "pkgs");
+                string packageSource = pathContext.PackageSource;
 
                 var prereleaseDependencyAName = "PreReleasePackageA";
                 var prereleaseDependencyAVersion = "4.8.0-beta00011";
@@ -6112,8 +6101,7 @@ namespace ClassLibrary
                 settings.AddNetStandardFeeds();
 
                 string testDirectory = pathContext.WorkingDirectory;
-                var mainDirectory = Directory.GetParent(testDirectory).FullName;
-                string packageSource = Path.Combine(mainDirectory, "pkgs");
+                string packageSource = pathContext.PackageSource;
 
                 var prereleaseDependencyAName = "PreReleasePackageA";
                 var prereleaseDependencyAVersion = "4.8.0-beta00011";
